Implement pipe IPC.
The message IPC is synchronous in 0.x which is going to make it unusable
in production so I've created a stream based `Conduit`. Instead of using
stream back pressure I'm using a queue that will pause on the round trip
to the child, from the request to the response with a default of 24
parallel requests, which is the same number I chose for HTTP server
Inlet. Doubtful that I'll ever use the round robin built into node to
delegate workers, or hand off sockets from a listener to workers.